Revert "Lookup instance availability for all AZs (#5390)"
This reverts commit 58be535592bd634a1f0d0637808bd008af1357f5.
