+++ pkg/apis/eksctl.io/v1alpha5/types.go
@@ -1268,9 +1268,6 @@ type NodePool interface {
 
 	// NGTaints returns the taints to apply for this nodegroup
 	NGTaints() []NodeGroupTaint
-
-	// InstanceTypeList returns a list of instances that are configured for that nodegroup
-	InstanceTypeList() []string
 }
 
 // VolumeMapping Additional Volume Configurations
