+++ pkg/ctl/cmdutils/cluster.go
@@ -44,3 +44,15 @@ func ToKubeNodeGroups(clusterConfig *api.ClusterConfig) []eks.KubeNodeGroup {
 	}
 	return kubeNodeGroups
 }
+
+// ToNodePools combines managed and self-managed nodegroups and returns a slice of api.NodePool
+func ToNodePools(clusterConfig *api.ClusterConfig) []api.NodePool {
+	var nodePools []api.NodePool
+	for _, ng := range clusterConfig.NodeGroups {
+		nodePools = append(nodePools, ng)
+	}
+	for _, ng := range clusterConfig.ManagedNodeGroups {
+		nodePools = append(nodePools, ng)
+	}
+	return nodePools
+}
